Begin rearrangement.
